+++ src/constants.js
@@ -3,10 +3,15 @@
 const MEDIA_TYPES = ["image", "video", "audio"];
 const SPECIAL_DIRS = {
   SEPARATOR: "---",
-  SOURCE_DOMAIN: ":sourcedomain:",
-  PAGE_DOMAIN: ":pagedomain:",
-  PAGE_URL: ":pageurl:",
-  DATE: ":date:"
+  SOURCE_DOMAIN: new RegExp(":sourcedomain:", "g"),
+  PAGE_DOMAIN: new RegExp(":pagedomain:", "g"),
+  PAGE_URL: new RegExp(":pageurl:", "g"),
+  DATE: new RegExp(":date:", "g"),
+  ISO8601_DATE: new RegExp(":isodate:", "g"),
+  UNIX_DATE: new RegExp(":unixdate:", "g"),
+  FILENAME: new RegExp(":filename:", "g"),
+  FILE_EXTENSION: new RegExp(":fileext:", "g"),
+  LINK_TEXT: new RegExp(":linktext:", "g")
 };
 
 // Export for testing