+++ src/options/options.js
@@ -4,7 +4,9 @@ const saveOptions = e => {
   browser.storage.local.set({
     links: document.querySelector("#links").checked,
     paths: document.querySelector("#paths").value.trim(),
+    filenamePatterns: document.querySelector("#filenamePatterns").value.trim(),
     prompt: document.querySelector("#prompt").checked,
+    promptIfNoExtension: document.querySelector("#promptIfNoExtension").checked,
     notifyOnSuccess: document.querySelector("#notifyOnSuccess").checked,
     notifyOnFailure: document.querySelector("#notifyOnFailure").checked,
     notifyDuration: document.querySelector("#notifyDuration").value
@@ -19,7 +21,9 @@ const restoreOptions = () => {
     .get([
       "links",
       "paths",
+      "filenamePatterns",
       "prompt",
+      "promptIfNoExtension",
       "notifyOnSuccess",
       "notifyOnFailure",
       "notifyDuration"
@@ -31,7 +35,15 @@ const restoreOptions = () => {
       document.querySelector("#prompt").checked =
         typeof result.prompt === "undefined" ? false : result.prompt;
 
+      document.querySelector("#promptIfNoExtension").checked =
+        typeof result.promptIfNoExtension === "undefined"
+          ? false
+          : result.promptIfNoExtension;
+
       document.querySelector("#paths").value = result.paths || ".";
+
+      document.querySelector("#filenamePatterns").value =
+        result.filenamePatterns || "";
 
       document.querySelector("#notifyOnSuccess").checked =
         typeof result.notifyOnSuccess === "undefined"
@@ -60,3 +72,9 @@ const addHelp = el => {
 document.addEventListener("DOMContentLoaded", restoreOptions);
 document.querySelector("#options").addEventListener("submit", saveOptions);
 document.querySelectorAll(".help").forEach(addHelp);
+
+if (browser === chrome) {
+  document.querySelectorAll(".chrome-only").forEach(el => {
+    el.classList.toggle("show");
+  });
+}